add fixed-width type aliases and convert the codebase to them
types.h defines i8-i64, u8-u64, usize, isize and b8 over stdint.h, with static asserts on u64 and pointer width. converts
declarations across all headers and sources.

findTerminator now returns isize rather than int, so an index into a usize-length buffer is no longer truncated on return

+++ include/server.h
@@ -1,6 +1,8 @@
 #ifndef SERVER_H
 #define SERVER_H
 
-int run_server(int port);
+#include "types.h"
+
+int run_server(u16 port);
 
 #endif
